Renamed constants and file names of navigators and screen names
