+++ src/constants/screen-names.ts
@@ -1,6 +1,8 @@
-export const NAVIGATOR_NAMES = {
+export const SCREEN_NAMES = {
   ACCOUNTS: 'ACCOUNTS_NAVIGATOR',
   GROUPS: 'GROUPS_NAVIGATOR',
+  HOME: 'HOME',
+  TRANSACTIONS_BY_ACCOUNT: 'TRANSACTIONS_BY_ACCOUNT',
 };
 
 export const ACCOUNTS_NAVIGATOR_SCREEN_NAMES = {
